requirements.txt
azure-ai-inference>=1.0.0b6
azure-identity>=1.15.0
httpx>=0.27.0
httpx-sse>=0.4.0
python-dotenv>=1.0.0
